Created a command line version
This version can now take a profile as input and write everything out to
a specified output directory.

I had an issue where the UML diagrams weren't getting generated.
Fixing this required removing `node_modules` and re-running
`yarn install`.

Frankly, at this point I'm starting to get JSON Schema fatigue
when specifying "shape" information.  I think JSON Schema is great
as a machine readable format, but the issue is writing it by hand.
On top of that, I find that I write my TS definitions first and the tool/IDE
support is much better there (despite leveraging $schema in my
JSON schema files).

As such, I think it would be useful to have a way to describe a profile
in TypeScript and then *generate* the profile from that.  So I need
to explore what makes sense there.  My goal isn't to require everybody
to use TypeScript.  But since all the information can be pretty easily
generated in TypeScript, one could easily create a kind of DSL in
TypeScript and then generate a profile from that.  This is a bit
of the reverse of what I was originally thinking about which had
JSON Schema generating TypeScript types.

Also got rid of a bunch of debugging output.

To run the CLI on the sample profile, just run `yarn start`.